Update polkadot-sdk to stable2506 (#76)
* update polkadot-sdk to stable2506

* clippy fixes

+++ pallets/author-slot-filter/src/migration.rs
@@ -75,7 +75,7 @@ where
 }
 
 fn percent_of_num(percent: Percent, num: u32) -> u32 {
-	percent.mul_ceil(num as u32)
+	percent.mul_ceil(num)
 }
 
 #[cfg(test)]
